feat(agents): single-page agent editor with live preview
Unify the agent create wizard and the per-section edit page into one
single-page editor (Gemini "Gem" style): a scrollable form column with
Gemini-minimal progressive disclosure next to a live preview pane, saved
with one button.

- New AgentEditor (shared by create + edit) and AgentPreview (static,
  live-updating preview with avatar, opening message and question pills).
- Core fields visible (icon+name, description, instructions, tools,
  knowledge); conversation and advanced settings in two collapsibles.
- Remove the max-tokens control from the UI (kept at its default).
- AgentSettingsPage becomes a thin loader; delete the 5-step wizard.
- Add formToPayload() to dedupe the create/edit payload mapping.
