+++ apps/web/src/features/agents/AgentBuilderPage.tsx
@@ -1,23 +1,17 @@
-import AgentBuilderForm from './AgentBuilderForm';
+import AgentEditor from './AgentEditor';
+import { EMPTY_FORM } from './agentFormState';
 
-import PageContainer from '@/components/common/PageContainer';
 import { useDocumentTitle } from '@/components/hooks/useDocumentTitle';
 
 /**
- * Route wrapper for the manual create entry point (`/agents/new/manual`). The
- * AI-assisted create renders {@link AgentBuilderForm} from the start-screen
- * orchestrator; editing uses {@link AgentSettingsPage}.
+ * Route wrapper for the manual create entry point (`/agents/new/manual`). Opens
+ * the single-page {@link AgentEditor} with an empty form. The AI-assisted create
+ * renders the same editor pre-filled from the start screen; editing reuses it too.
  */
 function AgentBuilderPage() {
   useDocumentTitle('Neuer Agent');
 
-  return (
-    <PageContainer maxWidth="md" title="Neuer Agent">
-      <div className="mx-auto max-w-3xl">
-        <AgentBuilderForm />
-      </div>
-    </PageContainer>
-  );
+  return <AgentEditor mode="create" initialState={EMPTY_FORM} />;
 }
 
 export default AgentBuilderPage;
